Technical/Update gem dependencies (#269)
* Updated gemspecs
* Updated codeclimate config
* Changed rspec helpers namespaces
* Updated reek config
